Don't throw away type annotations
The expression language defines a parser

    declParser :: BlinkParser (Name, Ty, Maybe SrcExp)

but this parser is only used at the very top-level (for MkProg). Everywhere
else we use a derived parser which throws away the type annotation if there
is a source expression:

    declParser' :: BlinkParser (Name, Either Ty SrcExp)
    declParser' = do
      (x, ty, mbinit) <- declParser
        return (x, case mbinit of Nothing -> Left ty ; Just ei -> Right ei)

I discussed this with Dimitri and we decided that it would be better not to
throw away these type annotations. This however implies some AST changes.
I've pushed this all the way through, but have marked a few TODOs that somebody
who is more familiar with the compiler should take a look at.