Descent LetERef in checkUnresolved
checkUnresolved descends into a term to find the main computation, skipping
over LetStruct, LetE, etc. But for some reason it was not skipping over
LetERef. Now that we've gotten rid of globals and make them LetERefs instead
this was causing some existing code that was accepted before to be rejected
with an "has unresolved type" error.