Allow omitted 'else' branch in every computation
Previously this was only permitted in "commands" -- that it, it was treated
in a similar way to a let without an `in` clause or bind instruction; however,
there is no reason for that so I've removed the special treatment of `if`
in commands and added support for omitted "else" clauses in the standard
definition of a computation.

The parse trees for all tests, including WiFi, are still precisely the same so
this does not affect existing code.

(Also a tiny bit of other cleanup in the parser.)

This closes #14.