Add support for local funs (front end only)
This commit adds support for local function definitions in the expression
language to the front end of the compiler: updated the parser, the ADT, the
renamer, and the type checker, but if you actually try to use local function
definitions the following functions will throw runtime exceptions:

- range analysis (`erange`)
- use-def analysis (`varUseDefs`)
- LUT analysis (`shouldLUT`)
- auto-LUT (`runAutoLUT`)
- code generation (`codeGenExp`)

This goes some way towards implementing #33.

NOTE: The AST for existing code is unaffected.