Fix and cleanup alength-elim
alength-elim replaces length(arr) expressions with the actual length of the
array; for arrays of polymorphic arrays, it replaced this by a term-level
variable that corresponded to the type-level length variable. This confusion of
type-level variables with term-level variables was not however supported in the
code generator, which would throw a runtime exception "Unbound variable in code
generation". This bug however never materialized due to a fortunate (or not so
fortunate, depending on your point of view) happenstance:

* Expression functions were almost always inlined, so that this situation never
  arose.
* The only place where they were _not_ inlined was when they appeared as
  arguments to map -- and in the comp language (including calls to 'map') we do
  not support polymorphism at all for some reason (#86).
* Likewise, for computation functions we do not support any kind of
  polymorphism.

I've changed it now so that it alength-elim does not apply to arrays with a
polymorphic width; the code generator deals correctly with such expressions.