Don't insert defaults into letrefs
The interpreter was picking defaults for letrefs, and leaving them inserted
even if we could not reduce the RHS to a value. We now don't do this anymore,
for performance reasons. This might be slgihtly less important if we fix
the representation of defaults in the AST too (by using SparseArray).