Fix bug in interpreter
We assumed that when we assigned to a variable with a currently-known-value
then we would have replaced that variable with its known value everywhere it
was used so far, and since we are now writing to it we don't need the old value
at all anymore. However, this is only true for assignments that replace the
entire value of the variable (`x := ..`), it is not true for assignments that
update the value of the variable (`x[0] = ..` or `x.foo = ..`).

Added test case for this specifically.